Revert "upgrade to skia 0.78.0 (#54)"
This reverts commit 12dea3692d6ac3bcc128e3dfb24e6a571207599c.

+++ skia-safe/src/core/font_arguments.rs
@@ -39,7 +39,7 @@ pub mod palette {
     #[derive(Copy, Clone, PartialEq, Eq, Default, Debug)]
     #[repr(C)]
     pub struct Override {
-        pub index: u16,
+        pub index: i32,
         pub color: Color,
     }
